src/requirements.txt
annotated-types==0.6.0
ansi2html==1.8.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.1.1
cycler==0.12.1
dacite==1.8.1
dash==2.14.1
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.13.1
Flask==3.0.0
fonttools==4.43.1
fqdn==1.5.1
fsspec==2023.10.0
h5py==3.10.0
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
imageio==2.32.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.26.0
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
multimethod==1.10
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nibabel==5.1.0
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.4
opencv-python==4.8.1.78
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandas-profiling==3.6.6
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
phik==0.12.3
pickleshare==0.7.5
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
plotly==5.18.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydicom==2.4.3
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.0
sympy==1.12
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
tensorboardX==2.6.2.2
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
torch==2.1.0
-e git+https://github.com/ncullen93/torchsample.git@bca89745bdcf2acb8cbe2feb1397e3fb94fa0540#egg=torchsample
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typeguard==4.1.5
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
visdom==0.2.4
visions==0.7.5
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wordcloud==1.9.2
ydata-profiling==4.6.1
zipp==3.17.0
